+++ R/mbrest_postprocess.R
@@ -1,12 +1,20 @@
-#' Post-process: derive predictions from optimization
+#' Postprocess mapbayr
 #'
+#' @name postprocess
 #' @inheritParams mbrest
 #' @param opt.value value returned by optimizer
-#' @param arg.optim,arg.ofv argument passed to optimizer
-#'
-#' @return a list of post processing values
+#' @param arg.optim,arg.ofv.fix,arg.ofv.id argument passed to optimizer
+#' @param post output of the post.process function
+#' @description Functions to generate postprocess after optimization.
+NULL
+#> NULL
+
+
+
+#' Post-process: derive predictions from optimization
+#' @rdname postprocess
 #' @export
-postprocess.optim <- function(x, data, opt.value, arg.optim, arg.ofv){
+postprocess.optim <- function(x, data, opt.value){
 
   final_eta <- opt.value[eta_names(x)] %>%
     as.double() %>%
@@ -56,13 +64,9 @@ postprocess.optim <- function(x, data, opt.value, arg.optim, arg.ofv){
 
 
 #' Post-process: Build the output (mbrests model object)
-#' @inheritParams mbrest
-#' @inheritParams postprocess.optim
-#' @param post output of the post.process function
-#'
-#' @return a mbrests model object
+#' @rdname postprocess
 #' @export
-postprocess.output <- function(x, data, arg.optim, arg.ofv, opt.value, post, output){
+postprocess.output <- function(x, arg.optim, arg.ofv.fix, arg.ofv.id, opt.value, post, output){
 
   if(!is.null(output)){
     if(output == "df") out <- map_dfr(post, "mapbay_tab")
@@ -70,9 +74,10 @@ postprocess.output <- function(x, data, arg.optim, arg.ofv, opt.value, post, out
   } else {
     out <- list(
       model = x,
-      data = bind_rows(unname(data)),
+      data = bind_rows(unname(map(arg.ofv.id, "data"))),
       arg.optim = arg.optim,
-      arg.ofv = arg.ofv,
+      arg.ofv.fix = arg.ofv.fix,
+      arg.ofv.id = arg.ofv.id,
       opt.value = map_dfr(opt.value, rownames_to_column, var = "method", .id = "ID"),
       final_eta = map(post, "final_eta"),
       mapbay_tab = map_dfr(post, "mapbay_tab")