+++ NEWS.md
@@ -1,6 +1,7 @@
 # mapbayr (Development version)
 
 - Data helpers are now 'mrgmod' methods
+- Fix bug #41
 
 # mapbayr 0.2.2