+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Data helpers are now 'mrgmod' methods
 - Maximum reset = 50
+- Maximum iteration defaults to 9999
+- New ini reset with samples in mvgauss.
 - Fix bugs #41 #42 #37
 
 # mapbayr 0.2.2